refactor(packages): getContractAddressforNetwork
