+++ packages/suite/src/utils/wallet/coinmarket/coinmarketUtils.ts
@@ -12,11 +12,8 @@ import TrezorConnect from '@trezor/connect';
 import regional from 'src/constants/wallet/coinmarket/regional';
 import { ExtendedMessageDescriptor, TrezorDevice } from 'src/types/suite';
 import { BuyTrade, SellFiatTrade, CryptoId } from 'invity-api';
-import {
-    DefinitionType,
-    getContractAddressForNetwork,
-    isTokenDefinitionKnown,
-} from '@suite-common/token-definitions';
+import { DefinitionType, isTokenDefinitionKnown } from '@suite-common/token-definitions';
+import { getContractAddressForNetwork } from '@suite-common/wallet-utils';
 import {
     CoinmarketAccountOptionsGroupOptionProps,
     CoinmarketAccountsOptionsGroupProps,
