+++ suite-common/token-definitions/src/tokenDefinitionsUtils.ts
@@ -1,5 +1,5 @@
 import { NetworkSymbol, getNetworkFeatures } from '@suite-common/wallet-config';
-import { parseAsset } from '@trezor/blockchain-link-utils/src/blockfrost';
+import { getContractAddressForNetwork } from '@suite-common/wallet-utils';
 
 import {
     DefinitionType,
@@ -8,27 +8,6 @@ import {
     TokenManagementAction,
     TokenManagementStorage,
 } from './tokenDefinitionsTypes';
-
-export const getContractAddressForNetwork = (
-    networkSymbol: NetworkSymbol,
-    contractAddress: string,
-) => {
-    switch (networkSymbol) {
-        case 'eth':
-            // Specyfing most common network as first case improves performance little bit
-            return contractAddress.toLowerCase();
-        case 'sol':
-        case 'dsol':
-            return contractAddress;
-        case 'ada':
-        case 'tada':
-            const { policyId } = parseAsset(contractAddress);
-
-            return policyId.toLowerCase();
-        default:
-            return contractAddress.toLowerCase();
-    }
-};
 
 // Using Set greatly improves performance of this function because of O(1) complexity instead of O(n) for Array.includes
 const tokenDefinitionsMap = new WeakMap<SimpleTokenStructure, Set<string>>();
