+++ suite-common/wallet-utils/src/__tests__/tokenUtils.test.ts
@@ -1,0 +1,13 @@
+import { getContractAddressForNetworkFixtures } from '../__fixtures__/tokenUtils';
+import { getContractAddressForNetwork } from '../tokenUtils';
+
+describe('getContractAddressForNetwork', () => {
+    getContractAddressForNetworkFixtures.forEach(
+        ({ testName, networkSymbol, contractAddress, expected }) => {
+            test(testName, () => {
+                const result = getContractAddressForNetwork(networkSymbol, contractAddress);
+                expect(result).toBe(expected);
+            });
+        },
+    );
+});
