+++ suite-common/wallet-utils/src/index.ts
@@ -20,5 +20,6 @@ export * from './antiFraud';
 export * from './stakingUtils';
 export * from './reviewTransactionUtils';
 export * from './filterReceiveAccounts';
+export * from './tokenUtils';
 
 export { analyzeTransactions as analyzeTransactionsFixtures } from './__fixtures__/transactionUtils';
